css/css-tables/tables-ng/table_minmax.html

<!doctype html>
<title>Table minmax tricks</title>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/check-layout-th.js"></script>
<script src="./support/table-inspector.js"></script>
<link rel="author" title="Aleks Totic" href="atotic@chromium.org" />
<link rel="help" href="https://www.w3.org/TR/css-tables-3/" />
<style>
   table {
    background: gray;
    border-spacing: 10px 10px;
    table-layout: auto;
  }
  td {
    padding: 0px;
    background: #BFB;
    font-size: 10px;
    box-sizing: border-box;
  }
  td > div {
    display: inline-block;
    background: rgba(56,162,56,0.3);
    line-height: 2;
  }
  .error {
    color: #b70000;
  }
  p {
    margin-top:4px;
    margin-bottom:4px;
  }
  body {
    counter-reset: testid;      /* Create a chapter counter scope */
  }
  .testdesc {
    white-space: pre-line;
    margin-top: 16px;
  }
  .testdesc:before {
    content: counter(testid) ". ";
    counter-increment: testid;  /* Add 1 to chapter */
  }
  .testdesc:first-line {
    font-family: monospace;
  }

  .container {
    border: 5px solid orange;
    margin-top: 8px;
    box-sizing: border-box;
    width: 600px;
  }
</style>
<h2>Table minmax</h2>

<p>How do different layout algorithms deal with table with an infinite max width?`</p>
<xmp>
  <table>
    <td style="width:100%"><div style="width:30px">100%</div></td>
    <td><div style="width:100px">100px</div></td>
  </table>
</xmp>
<p>Block layout clamps the max width, but  flex/grid/table layouts do not. What happens?</p>

<p class="testdesc">Block layout</p>
<div class="container">
  <table data-expected-width=590>
    <td style="width:100%"><div style="width:30px">100%</div></td>
    <td><div style="width:100px">100px</div></td>
  </table>
  <div>sibling</div>
</div>

<p class="testdesc">Block layout with floats</p>
<div class="container">
  What about this?
  <table style="float:left" data-expected-width=590>
    <td style="width:100%"><div style="width:30px">100%</div></td>
    <td><div style="width:100px">100px</div></td>
  </table>
  <div style="float:left">sibling</div>
  <div style="clear:both"></div>
</div>


<p class="testdesc">Table layout</p>

<table class="container" style="width:auto">
  <td>
    <table data-expected-width=160>
      <td style="width:100%"><div style="width:30px">100%</div></td>
      <td><div style="width:50px">50px</div><div style="width:50px">50px</div></td>
    </table>
  </td>
  <td><div style="width:100px;">sibling</div></td>
</table>

<p class="testdesc">Flex layout 1 1 auto</p>
<p class="error">FF grows table to the max flex available. Edge is bad, wider than
flexbox</p>
<div class="container" style="display:flex;justify-content:center">
  <table style="flex: 1 1 auto" data-expected-width=355>
    <td style="width:100%"><div style="width:30px">100%</div></td>
    <td><div style="width:50px">50px</div><div style="width:50px">50px</div></td>
  </table>
  <div style="background:yellow;flex:1 1 auto">100px</div>
</div>

<p class="testdesc">Flex layout 0 0 auto</p>

<div  class="container" style="display:flex;justify-content:center">
  <table style="flex: 0 0 auto" data-expected-width=160>
    <td style="width:100%"><div style="width:30px">100%</div></td>
    <td><div style="width:50px">50px</div><div style="width:50px">50px</div></td>
  </table>
  <div style="background:yellow;flex:1 1 auto">100px</div>
</div>

<script>
  checkLayout("table");
</script>
